move (xml)metadatacomplete tests
