--- jpa/eclipselink.jpa.test/src/it/assembly/eclipselink-xml-mapping-metadata-complete.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (c) 2019 Oracle and/or its affiliates. All rights reserved.
-
-    This program and the accompanying materials are made available under the
-    terms of the Eclipse Public License v. 2.0 which is available at
-    http://www.eclipse.org/legal/epl-2.0,
-    or the Eclipse Distribution License v. 1.0 which is available at
-    http://www.eclipse.org/org/documents/edl-v10.php.
-
-    SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
-
--->
-
-<component xmlns="http://maven.apache.org/ASSEMBLY-COMPONENT/2.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-           xsi:schemaLocation="http://maven.apache.org/ASSEMBLY-COMPONENT/2.0.0 http://maven.apache.org/xsd/assembly-component-2.0.0.xsd">
-    <fileSets>
-        <fileSet>
-            <directory>${project.build.testOutputDirectory}/${item}</directory>
-            <outputDirectory>.</outputDirectory>
-            <includes>
-                <include>org/eclipse/persistence/testing/models/jpa/xml/xmlmetadatacomplete/**</include>
-            </includes>
-        </fileSet>
-        <fileSet>
-            <directory>${project.build.testOutputDirectory}</directory>
-            <outputDirectory>.</outputDirectory>
-            <includes>
-                <include>org/eclipse/persistence/testing/models/jpa/xml/xmlmetadatacomplete/*</include>
-            </includes>
-        </fileSet>
-    </fileSets>
-</component>
