--- jpa/eclipselink.jpa.test/src/it/resources/eclipselink-xml-mapping-metadata-complete/ecl-xml-mapping-metadata-complete-mappings.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (c) 2018, 2020 Oracle and/or its affiliates. All rights reserved.
-
-    This program and the accompanying materials are made available under the
-    terms of the Eclipse Public License v. 2.0 which is available at
-    http://www.eclipse.org/legal/epl-2.0,
-    or the Eclipse Distribution License v. 1.0 which is available at
-    http://www.eclipse.org/org/documents/edl-v10.php.
-
-    SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
-
--->
-
-<entity-mappings version="1.0"
-    xmlns="http://java.sun.com/xml/ns/persistence/orm"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://java.sun.com/xml/ns/persistence/orm orm_1_0.xsd">
-    <persistence-unit-metadata>
-        <xml-mapping-metadata-complete/>
-    </persistence-unit-metadata>
-
-    <entity name="XMLOnlyEntity" class="org.eclipse.persistence.testing.models.jpa.xml.xmlmetadatacomplete.XMLOnlyEntity">
-        <attributes>
-            <id name="id"/>
-            <basic name="strData"/>
-        </attributes>
-    </entity>
-</entity-mappings>
