+++ jpa/eclipselink.jpa.testapps/jpa.test.performance/pom.xml
@@ -30,7 +30,7 @@
     <properties>
         <argLine/>
 <!--        <el.packager.fwk.exclusionFilter>%regex[.*TestRunner.*Bean.*]</el.packager.fwk.exclusionFilter>-->
-        <se.test.skip>false</se.test.skip>
+        <se.test.skip>true</se.test.skip>
         <server.test.skip>true</server.test.skip>
     </properties>
