+++ jpa/eclipselink.jpa.testapps/jpa.test.performance2/pom.xml
@@ -41,7 +41,7 @@
     <properties>
         <argLine/>
 <!--        <el.packager.fwk.exclusionFilter>%regex[.*TestRunner.*Bean.*]</el.packager.fwk.exclusionFilter>-->
-        <se.test.skip>false</se.test.skip>
+        <se.test.skip>true</se.test.skip>
         <server.test.skip>true</server.test.skip>
     </properties>
