+++ jpa/eclipselink.jpa.testapps/jpa.test.xml.metadatacomplete/src/main/java/org/eclipse/persistence/testing/models/jpa/xml/metadatacomplete/Hammer.java
