+++ jpa/eclipselink.jpa.testapps/jpa.test.xml.metadatacomplete/src/main/java/org/eclipse/persistence/testing/models/jpa/xml/metadatacomplete/Material.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright (c) 2011, 2022 Oracle and/or its affiliates. All rights reserved.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License v. 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0,
+ * or the Eclipse Distribution License v. 1.0 which is available at
+ * http://www.eclipse.org/org/documents/edl-v10.php.
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
+ */
+
+// Contributors:
+//     12/01/2010-2.2 Guy Pelletier
+//       - 331234: xml-mapping-metadata-complete overriden by metadata-complete specification
+package org.eclipse.persistence.testing.models.jpa.xml.metadatacomplete;
+
+import jakarta.persistence.Column;
+
+/**
+ * Included in the following mapping files:
+ * - eclipselink-xml-only-model/metadata-complete-mappings.
+ * - eclipselink-xml-only-model/xml-mapping-metadata-complete.
+ *
+ * Which are used in two separate persistence units from:
+ * - eclipselink-annotation-model/persistence.xml, namely:
+ *   - METADATA_COMPLETE
+ *   - XML_MAPPING_METADATA_COMPLETE
+ *
+ * @author gpelleti
+ */
+public class Material {
+    public int id;
+
+    @Column(name="MUCHO_EXPENSIVE")
+    public double cost;
+
+    public Material() {}
+}
