+++ jpa/eclipselink.jpa.testapps/jpa.test.xml.xmlmetadatacomplete/src/main/java/org/eclipse/persistence/testing/models/jpa/xml/xmlmetadatacomplete/XMLOnlyEntity.java
